feat: add commit specification and example to prompt
What
===

Extensive additions to `copilot-chat-commit-prompt`
with detailed specifications and examples of Conventional Commits and type descriptions.

Why
===

I added a description of the actual commit type, since the recognition of the commit type was quite inaccurate.
Since token consumption does not have to be so truncated these days, I also added an actual example.
I tried it for a while with my personal settings, and I think the accuracy has improved there.