fix: work on split JSON respose in on o1
I saw that there was no `DATA` format or anything like that,
and I implemented it assuming that o1 would never split JSON unlike the others.
That was not the case, so I fixed it.
I guess it might be due to the fact that I'm getting standard output with curl in the middle.
Since there is no identifier such as `DATA` in the o1 format,
after some thought I decided to use whether the process is alive or not as information.