fix: `copilot-chat-markdown-prompt` to empty
The default value for `copilot-chat-markdown-prompt` is an empty string, which is more useful than the current one.

Elements of the current prompt are not needed for their respective reasons.

* To specialize in coding functionality: it is tuned when it is built in as GitHub Copilot.
* To use Markdown formatting: it already does that by default.
* Using Neovim: actually using Emacs, which causes weird replies.

I actually set the prompt to `"Please respond in Japanese,`
convinced that this is more useful, I finally open a PR.
It might be better to set the language now with `copilot-chat-prompt-suffix` or something like that, which was added later.