feat: display the Copilot model name in response
The model name now appears in the buffer name,
but due to a refresh issue, it only shows the model when the buffer was created, causing me to be confused.
Also, when a model is changed after a response,
knowing which model the previous response was made with would reduce unnecessary queries.
Including the name of the model being used in the response header would solve these problems.
Either way, there is little overhead in displaying the information, since one line is used for the time display, etc.