Add optional custom prompt to copilot-chat-custom-prompt-selection
Allow users to provide a custom prompt directly to the function
copilot-chat-custom-prompt-selection. If no custom prompt is provided, the
function will read from the mini-buffer.

This makes it easy to for users to create customized functions for example:
```elisp
(defun copilot-chat-correct-typos ()
  (interactive)
  (copilot-chat-custom-prompt-selection
   "Please fix typos in the following text "))
```
Another enhancement is: if the region is inactive, use the whole buffer.