fix: move remove operation of old spiner overlay to chat buffer
The previous operation would have been done in the current buffer,
so if I focused on a different buffer and waited for Copilot Chat to respond,
the `Thinking` string would cover up a lot of the buffer.
Even though it disappears after the last focus,
it is hard to read and uncomfortable to work with the output on the side.